requirements.txt
pip install fastapi uvicorn[standard] httpx beautifulsoup4 httpx selenium
pip install fastapi[all] fastapi-users[sqlalchemy] python-jose[cryptography] passlib[bcrypt] email-validator
pip install celery[redis]

fastapi
uvicorn
sqlalchemy
databases
fastapi-users[sqlalchemy]
python-jose[cryptography]
passlib[bcrypt]
email-validator
alembic
celery[redis]
python-dotenv
aiosqlite
